Move passport config to common
